+++ flutter_google_places_sdk/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+* Upgraded dependencies versions. Now supporting null safety
+* Added sessionToken and origin params to findAutoCompletePredictions
+* Added fetchPlace with types.
+* Updated example for new params and fetchPlace request
+
 ## 0.1.0+3
 
 * Exposing Powered by google assets from platform interface